pyproject.toml
[tool.poetry]
name = "notif_service"
version = "0.1.0"
description = ""
authors = ["alexeygalt <alexeygalt@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0.1"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.9"
drf-spectacular = "^0.27.1"
djangorestframework = "^3.14.0"
django-filter = "^23.5"
djangorestframework-simplejwt = "^5.3.1"
google-auth = "^2.23.4"
facebook-sdk = "^3.1.0"
requests = "^2.31.0"
django-prometheus = "^2.3.1"
gunicorn = "^21.2.0"
gevent = "^23.9.1"
celery = "^5.3.6"
flower = "^2.0.1"
django-redis = "^5.4.0"
pytest = "^8.0.0"
pytest-django = "^4.8.0"
freezegun = "^1.4.0"
factory-boy = "^3.3.0"
pytest-cov = "^4.1.0"
pre-commit = "^3.6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
